separated appbar and sidebar components
