+++ src/components/navbar-App.tsx
@@ -1,96 +1,52 @@
 import { Link } from "react-router-dom";
-import { useUserContext } from "../context/userContext";
-import { AppBar, Box, Button, Grid, IconButton, Stack, Toolbar, Typography } from "@mui/material";
+import { styled } from "@mui/material/styles";
+import { Button, IconButton, Toolbar, Typography } from "@mui/material";
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
+
 import MenuIcon from "@mui/icons-material/Menu";
 
-// const drawerWidth = 240;
+import { useUserContext } from "../context/userContext";
 
-// const openedMixin = (theme: Theme): CSSObject => ({
-//   width: drawerWidth,
-//   transition: theme.transitions.create("width", {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.enteringScreen,
-//   }),
-//   overflowX: "hidden",
-// });
+const drawerWidth = 240;
 
-// const closedMixin = (theme: Theme): CSSObject => ({
-//   transition: theme.transitions.create("width", {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.leavingScreen,
-//   }),
-//   overflowX: "hidden",
-//   width: `calc(${theme.spacing(7)} + 1px)`,
-//   [theme.breakpoints.up("sm")]: {
-//     width: `calc(${theme.spacing(8)} + 1px)`,
-//   },
-// });
+interface AppBarProps extends MuiAppBarProps {
+  open?: boolean;
+}
 
-// const DrawerHeader = styled("div")(({ theme }) => ({
-//   display: "flex",
-//   alignItems: "center",
-//   justifyContent: "flex-end",
-//   padding: theme.spacing(0, 1),
-//   // necessary for content to be below app bar
-//   ...theme.mixins.toolbar,
-// }));
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== "open",
+})<AppBarProps>(({ theme, open }) => ({
+  zIndex: theme.zIndex.drawer + 1,
+  transition: theme.transitions.create(["width", "margin"], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
 
-// interface AppBarProps extends MuiAppBarProps {
-//   open?: boolean;
-// }
-
-// const AppBar = styled(MuiAppBar, {
-//   shouldForwardProp: (prop) => prop !== "open",
-// })<AppBarProps>(({ theme, open }) => ({
-//   zIndex: theme.zIndex.drawer + 1,
-//   transition: theme.transitions.create(["width", "margin"], {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.leavingScreen,
-//   }),
-//   ...(open && {
-//     marginLeft: drawerWidth,
-//     width: `calc(100% - ${drawerWidth}px)`,
-//     transition: theme.transitions.create(["width", "margin"], {
-//       easing: theme.transitions.easing.sharp,
-//       duration: theme.transitions.duration.enteringScreen,
-//     }),
-//   }),
-// }));
-
-// const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== "open" })(
-//   ({ theme, open }) => ({
-//     width: drawerWidth,
-//     flexShrink: 0,
-//     whiteSpace: "nowrap",
-//     boxSizing: "border-box",
-//     ...(open && {
-//       ...openedMixin(theme),
-//       "& .MuiDrawer-paper": openedMixin(theme),
-//     }),
-//     ...(!open && {
-//       ...closedMixin(theme),
-//       "& .MuiDrawer-paper": closedMixin(theme),
-//     }),
-//   })
-// );
-
-function AppNavbar() {
+function AppNavbar(props: any) {
   const { userProfile } = useUserContext();
+  const { open, setOpen } = props;
 
-  // const handleDrawerOpen = () => {
-  //   setOpen(true);
-  // };
-
-  // const handleDrawerClose = () => {
-  //   setOpen(false);
-  // };
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
 
   return (
     <AppBar
       position="fixed"
+      elevation={2}
       sx={{
         zIndex: (theme) => theme.zIndex.drawer + 1,
       }}
+      open={open}
     >
       <Toolbar
         sx={{
@@ -98,10 +54,9 @@ function AppNavbar() {
           flexDirection: "row",
           justifyContent: "space-between",
           backgroundColor: "#0047AB",
-          height: "8vh",
         }}
       >
-        {/* <IconButton
+        <IconButton
           color="inherit"
           aria-label="open drawer"
           onClick={handleDrawerOpen}
@@ -112,7 +67,7 @@ function AppNavbar() {
           }}
         >
           <MenuIcon />
-        </IconButton> */}
+        </IconButton>
         <Link
           to={"/app/dashboard"}
           style={{
